stop setting recurse => true for mongodb dbpath
The dbpath directory (/var/lib/mongodb) is configured in the mongodb
module with mode 0755, which is fine for a directory. Setting recurse =>
true causes that mode to be set for all of the subdirectories (also
fine) and contained files (not fine).

But the real problem is that some files are created when the service
starts with a mode like 0644. A Puppet run changes the mode on those
files and triggers a service restart, which causes the files to be
recreated with mode 0644, etc, etc.